requirements.txt
pdfplumber
faiss-cpu
numpy
openai
tiktoken
streamlit
